requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
bitarray==1.0.1
colorama==0.4.1
cycler==0.10.0
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
importlib-metadata==0.19
kiwisolver==1.1.0
matlabengineforpython===R2018a
matplotlib==3.1.1
monotonic==1.5
more-itertools==7.2.0
numpy==1.17.0
packaging==19.1
pandas==0.25.0
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
PyCRC==1.21
pyparsing==2.4.2
pyserial==3.4
pytest==5.0.1
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
six==1.12.0
tqdm==4.33.0
unireedsolomon==1.0
wcwidth==0.1.7
xlwt==1.3.0
zipp==0.5.2
